split swiftlint.rb file to some files
