+++ xcode/build_phases/multiple_swiftlint/array_extension.rb
@@ -1,0 +1,5 @@
+class Array
+    def nilOrEmpty?
+        self.nil? or self.empty?
+    end
+end
