+++ xcode/build_phases/multiple_swiftlint/command_utils.rb
@@ -1,0 +1,6 @@
+class CommandUtils
+    def self.make_command(command)
+        command = command.to_s
+        return `#{command}`
+    end
+end
